Reorganize headers
